# internal/delivery/docs/swagger.yaml
definitions:
...
  models.Note:
    properties:
      id:
        type: integer
      is_check_in:
        type: boolean
      place_id:
        type: integer
      properties: {}
      user_id:
        type: integer
    type: object
  models.NoteCreate:
    properties:
      place_id:
        type: integer
      properties: {}
      user_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /note/by_user_and_place_ids:
    get:
      consumes:
      - application/json
      parameters:
      - description: user_id
        in: query
        name: user_id
        required: true
        type: integer
      - description: place_id
        in: query
        name: place_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully
          schema:
            $ref: '#/definitions/models.Note'
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - note
  /note/by_user_id:
    get:
      consumes:
      - application/json
      parameters:
      - description: User id
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully
          schema:
            items:
              $ref: '#/definitions/models.Note'
            type: array
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - note
  /note/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Note
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.NoteCreate'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created note with id
          schema:
            type: integer
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      tags:
      - note
...
swagger: "2.0"